requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
easyocr==1.7.1
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.3
frozenlist==1.4.0
idna==3.4
imageio==2.31.3
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
networkx==3.1
ninja==1.11.1
numpy==1.25.2
openai==0.28.0
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-bidi==0.4.2
python-dotenv==1.0.0
pytweening==1.0.7
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.8.30
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
